+++ custom_components/duofern/switch.py
@@ -811,7 +811,7 @@ class DuoFernSwitch(CoordinatorEntity[DuoFernCoordinator], SwitchEntity):
             model=self._device_code.device_type_name,
             serial_number=self._hex_code,
             sw_version=state.status.version if state else None,
-            via_device=(DOMAIN, self.coordinator.system_code.hex),
+            via_device_id=self.coordinator.stick_device_id,
         )
 
     @callback
@@ -819,7 +819,9 @@ class DuoFernSwitch(CoordinatorEntity[DuoFernCoordinator], SwitchEntity):
         state = self._device_state
         if state and state.status.version:
             device_reg = dr.async_get(self.hass)
-            device = device_reg.async_get_device(identifiers={(DOMAIN, self._hex_code)})
+            device = device_reg.async_get_device_by_identifier(
+                (DOMAIN, self._hex_code), self.coordinator.config_entry.entry_id
+            )
             if device and device.sw_version != state.status.version:
                 device_reg.async_update_device(
                     device.id, sw_version=state.status.version
@@ -975,7 +977,9 @@ class DuoFernAutomationSwitch(
         state = self._device_state
         if state and state.status.version:
             device_reg = dr.async_get(self.hass)
-            device = device_reg.async_get_device(identifiers={(DOMAIN, self._hex_code)})
+            device = device_reg.async_get_device_by_identifier(
+                (DOMAIN, self._hex_code), self.coordinator.config_entry.entry_id
+            )
             if device and device.sw_version != state.status.version:
                 device_reg.async_update_device(
                     device.id, sw_version=state.status.version